Linting update
